refactor : 전체적인 구조 및 css 재작업
