+++ src/components/ListItems.jsx
@@ -3,18 +3,7 @@ import "./../css/ListItems.css";
 
 function ListItems({ items, gridCol }) {
   const className = `ListItems ${gridCol}`;
-  return (
-    <ul className={className}>
-      {items &&
-        items.map((item) => {
-          return (
-            <li key={item.id}>
-              <Card item={item} />
-            </li>
-          );
-        })}
-    </ul>
-  );
+  return;
 }
 
 export default ListItems;
