+++ src/css/Header.css
@@ -13,15 +13,15 @@
 .logo_words {
   display: none;
 }
-nav {
+.header__nav {
   width: 90%;
   max-width: 1520px;
   min-width: 343px;
-  margin: 0 auto;
   display: flex;
   justify-content: space-between;
   align-items: center;
   font-weight: 700;
+  margin: 0 auto;
 }
 .nav__ul {
   margin-right: auto;
