+++ src/css/Item.css
@@ -1,37 +1,50 @@
-.Item {
+.Item .lists {
   margin: 94px auto 0;
-  width: 1201px;
   display: flex;
   flex-direction: column;
   gap: 40px;
+  max-width: 1201px;
+  width: 95%;
+  min-width: 800px;
+  padding: 15px;
 }
 
-.Item section {
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
+.Item .lists .ListSection .List {
+  display: grid;
+  gap: 16px;
 }
 
-.Item .title {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
+.Item .lists .ListSection:first-child .List {
+  grid-template-columns: repeat(4, 1fr);
+}
+.Item .lists .ListSection:nth-child(2) .List {
+  grid-template-columns: repeat(5, 1fr);
 }
 
-.Item .OrderNav {
-  display: flex;
-  gap: 12px;
+@media (max-width: 744px) {
+  .Item .lists {
+    max-width: 696px;
+    width: 95%;
+    min-width: 470px;
+  }
+  .Item .lists .ListSection:first-child .List {
+    gap: 24px;
+    grid-template-columns: repeat(2, 1fr);
+  }
+  .Item .lists .ListSection:nth-child(2) .List {
+    grid-template-columns: repeat(3, 1fr);
+  }
 }
-
-.Item .OrderNav input[type="text"] {
-  padding: 9px 16px;
-  background-color: var(--gray100);
-  width: 325px;
-  border-radius: 12px;
-  font-size: 1.6rem;
-  font-family: var(--main-font);
-}
-
-.Item .OrderNav input[type="text"]::placeholder {
-  color: var(--gray400);
+@media (max-width: 375px) {
+  .Item .lists {
+    width: 344px;
+    min-width: 344px;
+  }
+  .Item .lists .ListSection:first-child .List {
+    grid-template-columns: 1fr;
+  }
+  .Item .lists .ListSection:nth-child(2) .List {
+    grid-template-columns: repeat(2, 1fr);
+    gap: 8px;
+  }
 }
