--- src/css/List.css
@@ -1,72 +1,0 @@
-.List {
-  width: 1201px;
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-  margin: 0 auto;
-}
-.List__head {
-  display: flex;
-  width: 100%;
-  position: relative;
-}
-.List__title {
-  font-size: 2rem;
-  margin-right: auto;
-  display: inline;
-}
-
-.OrderNav__bottom {
-  display: flex;
-  gap: 12px;
-}
-
-.OrderNav button {
-  background-color: var(--main);
-  font-family: var(--main-font);
-  color: var(--white);
-  padding: 11px 23px;
-  border-radius: 8px;
-}
-.OrderNav select {
-  width: 130px;
-  border: 1px solid var(--gray200);
-  border-radius: 12px;
-}
-@media (max-width: 1350px) {
-  .List {
-    width: 95%;
-  }
-}
-@media (max-width: 1199px) {
-  .List {
-    width: 696px;
-  }
-}
-@media (max-width: 767px) {
-  .List {
-    width: 344px;
-  }
-  .List__head:has(.OrderNav) {
-    margin-bottom: 50px;
-  }
-  .List__head:has(.OrderNav) .List__title {
-    position: absolute;
-    top: 7px;
-  }
-  .OrderNav {
-    justify-content: end;
-    width: max-content;
-    flex-wrap: wrap;
-  }
-
-  .OrderNav input[type="text"] {
-    width: auto;
-    order: 2;
-  }
-
-  .OrderNav select {
-    order: 3;
-    width: 42px;
-  }
-}
