+++ src/css/ListSection.css
@@ -1,0 +1,6 @@
+.ListSection {
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+  width: 100%;
+}
