+++ src/pages/Main.jsx
@@ -1,7 +1,13 @@
 import React from "react";
 
+import Header from "../components/Header";
+
 function Main() {
-  return <div>Main</div>;
+  return (
+    <div className="Main">
+      <Header />
+    </div>
+  );
 }
 
 export default Main;
